added discount strategy, comments & refactor

+++ src/main/java/com/unishop/marketplace/models/CartItem.java
@@ -1,7 +1,7 @@
 package com.unishop.marketplace.models;
 
-public record CartItem(Product product, long quantity) {
-    public CartItem updatedQuantity(long quantity){
+public record CartItem(Product product, int quantity) {
+    public CartItem updatedQuantity(int quantity){
         return new CartItem(this.product, quantity);
     }
 }
